requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.15.1
anyio==3.6.2
apache-airflow==2.10.5
apache-airflow-providers-common-compat==1.5.1
apache-airflow-providers-common-io==1.5.1
apache-airflow-providers-common-sql==1.12.0
apache-airflow-providers-fab==1.5.3
apache-airflow-providers-ftp==3.8.0
apache-airflow-providers-http==4.9.0
apache-airflow-providers-imap==3.5.0
apache-airflow-providers-mysql==5.5.4
apache-airflow-providers-smtp==2.0.1
apache-airflow-providers-sqlite==3.7.1
apache-airflow-providers-telegram==4.5.2
apispec==6.8.1
argcomplete==2.0.5
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
Babel==2.12.1
blinker==1.9.0
cachelib==0.9.0
cattrs==22.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
clickclick==20.10.2
colorama==0.4.6
colorlog==6.9.0
ConfigUpdater==3.1.1
connexion==2.14.2
cron_descriptor==1.2.35
croniter==6.0.0
cryptography==44.0.2
Deprecated==1.2.13
dill==0.3.1.1
dnspython==2.3.0
docutils==0.19
email-validator==1.3.1
Flask==2.2.3
Flask-AppBuilder==4.5.3
Flask-Babel==2.0.0
Flask-Caching==2.0.2
Flask-JWT-Extended==4.4.4
Flask-Limiter==3.3.0
Flask-Login==0.6.2
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.1.1
frozenlist==1.3.3
fsspec==2025.3.0
google-re2==1.1.20240702
googleapis-common-protos==1.69.2
graphviz==0.20.1
grpcio==1.71.0
gunicorn==20.1.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.4
importlib_metadata==8.6.1
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
lazy-object-proxy==1.9.0
limits==3.2.0
linkify-it-py==2.0.0
lockfile==0.12.2
Mako==1.2.4
Markdown==3.4.1
markdown-it-py==2.2.0
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
mdit-py-plugins==0.3.5
mdurl==0.1.2
methodtools==0.4.7
more-itertools==9.1.0
multidict==6.0.4
mysql-connector-python==9.2.0
mysqlclient==2.2.7
numpy==2.3.0
opentelemetry-api==1.31.0
opentelemetry-exporter-otlp==1.31.0
opentelemetry-exporter-otlp-proto-common==1.31.0
opentelemetry-exporter-otlp-proto-grpc==1.31.0
opentelemetry-exporter-otlp-proto-http==1.31.0
opentelemetry-proto==1.31.0
opentelemetry-sdk==1.31.0
opentelemetry-semantic-conventions==0.52b0
ordered-set==4.1.0
packaging==23.2
pandas==2.3.0
pathspec==0.9.0
pendulum==2.1.2
pluggy==1.5.0
prison==0.2.1
protobuf==5.29.3
psutil==5.9.4
pycparser==2.21
pydantic==1.10.21
Pygments==2.14.0
PyJWT==2.6.0
PyMySQL==1.1.1
pyparsing==3.0.9
pyrsistent==0.19.3
python-daemon==3.0.0
python-dateutil==2.8.2
python-nvd3==0.15.0
python-slugify==8.0.1
python-telegram-bot==22.0
pytz==2022.7.1
pytzdata==2020.1
PyYAML==6.0
referencing==0.36.2
requests==2.28.2
requests-toolbelt==0.10.1
rfc3339-validator==0.1.4
rfc3986==1.5.0
rich==13.3.2
rich_argparse==1.0.0
rpds-py==0.23.1
scikit-learn==1.7.0
scipy==1.15.3
setproctitle==1.3.5
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.46
SQLAlchemy-JSONField==1.0.1.post0
SQLAlchemy-Utils==0.40.0
sqlparse==0.4.3
tabulate==0.9.0
tenacity==8.2.2
termcolor==2.2.0
text-unidecode==1.3
threadpoolctl==3.6.0
typing_extensions==4.12.2
tzdata==2025.2
uc-micro-py==1.0.1
unicodecsv==0.14.1
universal_pathlib==0.2.6
urllib3==1.26.14
Werkzeug==2.2.3
wirerope==1.0.0
wrapt==1.15.0
WTForms==3.0.1
yarl==1.8.2
zipp==3.21.0
